+++ app/controllers/tasks/biological_associations/biological_associations_graph_controller.rb
@@ -1,0 +1,4 @@
+class Tasks::BiologicalAssociations::BiologicalAssociationsGraphController < ApplicationController
+  include TaskControllerConfiguration
+
+end