+++ app/controllers/taxon_name_classifications_controller.rb
@@ -1,7 +1,7 @@
 class TaxonNameClassificationsController < ApplicationController
   include DataControllerConfiguration::ProjectDataControllerConfiguration
 
-  before_action :set_taxon_name_classification, only: [:update, :destroy, :show]
+  before_action :set_taxon_name_classification, only: [:update, :destroy, :show, :api_show, :edit]
   after_action -> { set_pagination_headers(:taxon_name_classifications) }, only: [:index, :api_index], if: :json_request?
 
   # GET /taxon_name_classifications
@@ -22,8 +22,12 @@ class TaxonNameClassificationsController < ApplicationController
     end
   end
 
-  # GET /taxon_name_classicfications/id.json
+  # GET /taxon_name_classifications/id.json
   def show
+  end
+
+  def edit
+    redirect_to new_taxon_name_task_path(taxon_name_id: @taxon_name_classification.taxon_name)
   end
 
   # POST /taxon_name_classifications
@@ -33,12 +37,16 @@ class TaxonNameClassificationsController < ApplicationController
 
     respond_to do |format|
       if @taxon_name_classification.save
-        format.html { redirect_to url_for(@taxon_name_classification.taxon_name.metamorphosize),
-                      notice: 'Taxon name classification was successfully created.' }
+        format.html {
+          redirect_to url_for(@taxon_name_classification.taxon_name.metamorphosize),
+                      notice: 'Taxon name classification was successfully created.'
+        }
         format.json { render :show, status: :created, location: @taxon_name_classification.metamorphosize }
       else
-        format.html { redirect_back(fallback_location: (request.referer || root_path),
-                                    notice:            'Taxon name classification was NOT successfully created.') }
+        format.html {
+          redirect_back(fallback_location: (request.referer || root_path),
+                        notice: 'Taxon name classification was NOT successfully created.')
+        }
         format.json { render json: @taxon_name_classification.errors, status: :unprocessable_entity }
       end
     end
@@ -50,12 +58,16 @@ class TaxonNameClassificationsController < ApplicationController
     respond_to do |format|
       if @taxon_name_classification.update(taxon_name_classification_params)
         @taxon_name_classification = TaxonNameClassification.find(@taxon_name_classification.id)
-        format.html { redirect_back(fallback_location: (request.referer || root_path),
-                                    notice:            'Taxon name classification was successfully updated.') }
+        format.html {
+          redirect_back(fallback_location: (request.referer || root_path),
+                        notice: 'Taxon name classification was successfully updated.')
+        }
         format.json { render :show, status: :ok, location: @taxon_name_classification.metamorphosize }
       else
-        format.html { redirect_back(fallback_location: (request.referer || root_path),
-                                    notice:            'Taxon name classification was NOT successfully updated.') }
+        format.html {
+          redirect_back(fallback_location: (request.referer || root_path),
+                        notice: 'Taxon name classification was NOT successfully updated.')
+        }
         format.json { render json: @taxon_name_classification.errors, status: :unprocessable_entity }
       end
     end
@@ -64,13 +76,15 @@ class TaxonNameClassificationsController < ApplicationController
   def new
     taxon_name_id = params[:taxon_name_classification].try(:[], :taxon_name_id)
     taxon_name = TaxonName.where(
-      id:         taxon_name_id,
-      project_id: sessions_current_project_id
+      id: taxon_name_id,
+      project_id: sessions_current_project_id,
     ).first
 
-    @taxon_name_classification = TaxonNameClassification.new(
-      taxon_name: taxon_name ? taxon_name : TaxonName.new()
-    )
+    if taxon_name
+      redirect_to new_taxon_name_task_path(taxon_name_id: taxon_name.id)
+    else
+      redirect_to new_taxon_name_task_path, notice: 'Create TaxonNameClassifications in the context of a TaxonName.'
+    end
   end
 
   # DELETE /taxon_name_classifications/1
@@ -80,7 +94,7 @@ class TaxonNameClassificationsController < ApplicationController
     respond_to do |format|
       if @taxon_name_classification.destroyed?
         format.html { destroy_redirect @taxon_name_classification, notice: 'Taxon name classification was successfully destroyed.' }
-        format.json { head :no_content}
+        format.json { head :no_content }
       else
         format.html { destroy_redirect @taxon_name_classification, notice: 'Taxon name classification was not destroyed, ' + @taxon_name_classification.errors.full_messages.join('; ') }
         format.json { render json: @taxon_name_classification.errors, status: :unprocessable_entity }
@@ -107,15 +121,15 @@ class TaxonNameClassificationsController < ApplicationController
   def autocomplete
     # TODO: make reasonable, if used
     @taxon_name_classifications = TaxonNameClassification.where(id: params[:term]).with_project_id(params[:project_id])
-    data                        = @taxon_name_classifications.collect do |t|
-      {id:              t.id,
-       label:           TaxonNameClassificationsHelper.taxon_name_classification_tag(t),
+    data = @taxon_name_classifications.collect do |t|
+      { id: t.id,
+       label: TaxonNameClassificationsHelper.taxon_name_classification_tag(t),
        response_values: {
-         params[:method] => t.id
-       },
-       label_html:      TaxonNameClassificationsHelper.taxon_name_classification_tag(t)
-       #  render_to_string(:partial => 'shared/autocomplete/taxon_name.html', :object => t)
-      }
+        params[:method] => t.id,
+      },
+       label_html: TaxonNameClassificationsHelper.taxon_name_classification_tag(t)
+ #  render_to_string(:partial => 'shared/autocomplete/taxon_name.html', :object => t)
+         }
     end
 
     render json: data
@@ -123,8 +137,8 @@ class TaxonNameClassificationsController < ApplicationController
 
   # GET /taxon_name_classifications/download
   def download
-    send_data Export::Download.generate_csv(TaxonNameClassification.where(project_id: sessions_current_project_id)),
-      type: 'text', filename: "taxon_name_classifications_#{DateTime.now}.csv"
+    _data Export::Download.generate_csv(TaxonNameClassification.where(project_id: sessions_current_project_id)),
+          type: 'text', filename: "taxon_name_classifications_#{DateTime.now}.csv"
   end
 
   def taxon_name_classification_types
@@ -155,6 +169,7 @@ class TaxonNameClassificationsController < ApplicationController
   def taxon_name_classification_params
     params.require(:taxon_name_classification).permit(
       :taxon_name_id, :type,
-      origin_citation_attributes: [:id, :_destroy, :source_id, :pages])
+      origin_citation_attributes: [:id, :_destroy, :source_id, :pages],
+    )
   end
 end