+++ app/javascript/vue/components/Form/FormCitation.vue
@@ -170,6 +170,7 @@ watch(sourceId, async (newId, oldId) => {
   if (newId) {
     if (newId !== oldId && newId !== source.value?.id) {
       source.value = (await Source.find(newId)).body
+      citation.value._label = source.value.cached
     }
   } else {
     source.value = undefined
@@ -189,6 +190,7 @@ function setSource(value) {
   source.value = value
   sessionStorage.setItem(STORAGE.sourceId, value.id)
   citation.value.source_id = value.id
+  citation.value._label = value.cached
 
   emit('source', value)
 }