+++ app/javascript/vue/components/radials/linker/links/BiologicalAssociation.js
@@ -1,4 +1,8 @@
 export const BiologicalAssociation = {
-  all: [],
-  ids: []
+  all: [
+    // TODO: link to new biological association task
+  ],
+  ids: [
+    // TODO: link to new biological association task
+  ]
 }