+++ app/javascript/vue/tasks/collection_objects/filter/constants/layouts.js
@@ -171,6 +171,17 @@ export const LAYOUTS = {
     includes: {}
   },
 
+  Deaccessioned: {
+    properties: {
+      collection_object: [
+        'accessioned_at',
+        'deaccessioned_at',
+        'deaccession_reason',
+      ],
+    },
+    includes: {}
+  },
+
   Paleo: {
     properties: {
       collection_object: [